+++ src/modules/contact/contact.service.ts
@@ -1,13 +1,40 @@
-import { Injectable } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateContactDto } from './dto/create-contact.dto';
 import { UpdateContactDto } from './dto/update-contact.dto';
 import { ContactRepository } from './repositories/contact.repository';
+import { ClientRepository } from '../client/repositories/client.repository';
 
 @Injectable()
 export class ContactService {
-  constructor(private contactRepository: ContactRepository) {}
+  constructor(
+    private contactRepository: ContactRepository,
+    private clientRepository: ClientRepository
+  ) {}
 
   async create(id: string, createContactDto: CreateContactDto) {
+    const findClient = await this.clientRepository.findOne(id);
+    if (!findClient) {
+      throw new NotFoundException('Cliente não encontrado');
+    }
+
+    const findContactEmail = await this.contactRepository.findOneByEmail(
+      createContactDto.email
+    );
+    if (findContactEmail) {
+      throw new ConflictException('Contato com este email já cadastrado!');
+    }
+
+    const findContactPhone = await this.contactRepository.findOneByPhone(
+      createContactDto.phone
+    );
+    if (findContactPhone) {
+      throw new ConflictException('Contato com este telefone já cadastrado!');
+    }
+
     const contact = await this.contactRepository.create(id, createContactDto);
     return contact;
   }
@@ -18,10 +45,18 @@ export class ContactService {
 
   async findOne(id: string) {
     const findContact = await this.contactRepository.findOne(id);
+    if (!findContact) {
+      throw new NotFoundException('Contato não encontrado');
+    }
     return findContact;
   }
 
   async update(id: string, updateContactDto: UpdateContactDto) {
+    const findContact = await this.contactRepository.findOne(id);
+    if (!findContact) {
+      throw new NotFoundException('Contato não encontrado');
+    }
+
     const updatedContact = await this.contactRepository.update(
       id,
       updateContactDto
@@ -30,6 +65,11 @@ export class ContactService {
   }
 
   async remove(id: string) {
+    const findContact = await this.contactRepository.findOne(id);
+    if (!findContact) {
+      throw new NotFoundException('Contato não encontrado');
+    }
+
     return await this.contactRepository.delete(id);
   }
 }
